Revert LcmSubscriber header/source file separation
Revert a0b1d74091ed316d4b7e7abb61070ece3161ebc9.  The LcmSubscriber implementation requires usage of the CTRL_C_DETECTED global variable to indicate stopping of the realtime thread.  CTRL_C_DETECTED is defined as static, and thus limited to file scope.  The header/source files should be separated once the scoping issues with CTRL_C_DETECTED are fixed.
