+++ src/CMakeLists.txt
@@ -30,7 +30,6 @@ set(SOURCE_LIST "pi3hat.cpp"
                 "robot_base.cpp"
                 "joint_base.cpp"
                 "cartesian_leg.cpp"
-                "lcm_subscriber.cpp"
         )
 
 # Make an automatic library - will be static or dynamic based on user setting
